+++ sdk-lib/src/androidTest/java/cash/z/ecc/android/sdk/internal/transaction/PersistentTransactionManagerTest.kt
@@ -10,19 +10,14 @@ import cash.z.ecc.android.sdk.db.entity.isCancelled
 import cash.z.ecc.android.sdk.internal.TroubleshootingTwig
 import cash.z.ecc.android.sdk.internal.Twig
 import cash.z.ecc.android.sdk.internal.service.LightWalletService
-import cash.z.ecc.android.sdk.internal.twig
+import cash.z.ecc.android.sdk.model.Account
 import cash.z.ecc.android.sdk.model.Zatoshi
 import cash.z.ecc.android.sdk.test.ScopedTest
 import cash.z.ecc.fixture.DatabaseNameFixture
 import cash.z.ecc.fixture.DatabasePathFixture
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.stub
-import kotlinx.coroutines.cancel
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.drop
-import kotlinx.coroutines.flow.launchIn
-import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import org.junit.Before
 import org.junit.BeforeClass
@@ -72,7 +67,7 @@ class PersistentTransactionManagerTest : ScopedTest() {
         MockitoAnnotations.openMocks(this)
         mockEncoder.stub {
             onBlocking {
-                createTransaction(any(), any(), any(), any(), any())
+                createTransaction(any(), any(), any(), any())
             }.thenAnswer {
                 runBlocking {
                     delay(200)
@@ -83,32 +78,8 @@ class PersistentTransactionManagerTest : ScopedTest() {
     }
 
     @Test
-    fun testCancellation_RaceCondition() = runBlocking {
-        val tx = manager.initSpend(Zatoshi(1234), "taddr", "memo-good", 0)
-        val txFlow = manager.monitorById(tx.id)
-
-        // encode TX
-        testScope.launch {
-            twig("ENCODE: start"); manager.encode("fookey", tx); twig("ENCODE: end")
-        }
-
-        // then cancel it before it is done encoding
-        testScope.launch {
-            delay(100)
-            twig("CANCEL: start"); manager.cancel(tx.id); twig("CANCEL: end")
-        }
-
-        txFlow.drop(2).onEach {
-            twig("found tx: $it")
-            assertTrue(it.isCancelled(), "Expected the encoded tx to be cancelled but it wasn't")
-            twig("found it to be successfully cancelled")
-            testScope.cancel()
-        }.launchIn(testScope).join()
-    }
-
-    @Test
     fun testCancel() = runBlocking {
-        var tx = manager.initSpend(Zatoshi(1234), "a", "b", 0)
+        var tx = manager.initSpend(Zatoshi(1234), "a", "b", Account.DEFAULT)
         assertFalse(tx.isCancelled())
         manager.cancel(tx.id)
         tx = manager.findById(tx.id)!!
@@ -117,7 +88,7 @@ class PersistentTransactionManagerTest : ScopedTest() {
 
     @Test
     fun testAbort() = runBlocking {
-        var tx: PendingTransaction? = manager.initSpend(Zatoshi(1234), "a", "b", 0)
+        var tx: PendingTransaction? = manager.initSpend(Zatoshi(1234), "a", "b", Account.DEFAULT)
         assertNotNull(tx)
         manager.abort(tx)
         tx = manager.findById(tx.id)